+++ bvp/api/v1/implementations.py
@@ -1,7 +1,7 @@
 import isodate
-from typing import Tuple, Union
+from typing import List, Tuple, Union
+from datetime import datetime as datetime_type, timedelta
 
-from flask import request
 from flask_json import as_json
 from flask_security import current_user
 from isodate import parse_duration
@@ -15,11 +15,18 @@ from bvp.api.common.responses import (
     request_processed,
 )
 from bvp.data.services.resources import get_assets
-from bvp.api.common.utils.api_utils import parse_entity_address, update_beliefs
+from bvp.api.common.utils.api_utils import (
+    parse_entity_address,
+    update_beliefs,
+    groups_to_dict,
+)
 from bvp.api.common.utils.validators import (
     type_accepted,
     units_accepted,
     connections_required,
+    optional_sources_accepted,
+    optional_resolutions_accepted,
+    optional_horizon_accepted,
     period_required,
     values_required,
     validate_entity_address,
@@ -28,12 +35,22 @@ from bvp.api.common.utils.validators import (
 
 @type_accepted("GetMeterDataRequest")
 @units_accepted("MW")
+@optional_resolutions_accepted("PT15M")
 @connections_required
+@optional_sources_accepted(preferred_source="MDC")
+@optional_horizon_accepted("-PT15M")
 @period_required
 @as_json
 def get_meter_data_response(
-    unit, connection_groups, start, duration
-) -> Union[dict, Tuple[dict, int]]:
+    unit,
+    resolution,
+    connection_groups,
+    horizon,
+    start,
+    duration,
+    preferred_source_ids,
+    fallback_source_ids,
+) -> Tuple[dict, int]:
     """
     Use marshmallow to connect SQLAlchemy-modelled data to the outside world.
     Only supports GET requests.
@@ -43,101 +60,17 @@ def get_meter_data_response(
     In all cases, the API defaults to use shorthand for univariate timeseries data,
     in which the data resolution can be derived by dividing the duration of the time window over the number of values.
     """
-    print(request)
-    from flask import current_app
 
-    current_app.logger.info("GETTING")
-
-    # Retrieve meter data from database
-    connections = connection_groups[
-        0
-    ]  # only one connection group allowed for get_meter_data
-    group = []
-    for asset in get_assets():
-        if asset.name in connections or str(asset.id) in connections:
-            print(
-                "For user %s, I chose Asset %s, id: %s"
-                % (current_user, asset, asset.id)
-            )
-            # TODO: use bvp.data.services.get_power?
-            # Maybe make it possible to return the actual DB objects we want here.
-            measurement_frequency = isodate.parse_duration(
-                "PT15M"
-            )  # Todo: actually look up the measurement frequency
-            measurements = Power.query.filter(
-                (Power.datetime > start - measurement_frequency)
-                & (Power.datetime < start + duration)
-                & (Power.asset_id == asset.id)
-            ).all()
-            # Todo: Check if all the data is there, otherwise return what is known and warn the user
-            start_buffer = (
-                measurements[0].datetime + measurement_frequency - start
-            ) % measurement_frequency
-            end_buffer = (
-                start + duration - measurements[-1].datetime
-            ) % measurement_frequency
-            values = [measurement.value for measurement in measurements]
-            # datetimes = [isodate.datetime_isoformat(measurement.datetime) for measurement in measurements]
-            # print(datetimes)
-            # print(len(values))
-            # TODO: convert to requested unit
-
-            # Start building a dictionary with the response
-            if start_buffer or end_buffer:  # timeseries is component-wise univariate
-                timeseries = []
-                if start_buffer:
-                    timeseries.append(
-                        dict(
-                            value=values[0],
-                            start=isodate.datetime_isoformat(start),
-                            duration=isodate.duration_isoformat(start_buffer),
-                        )
-                    )
-                    del values[0]
-                if end_buffer:
-                    temp_value = values[-1]
-                    del values[-1]
-                timeseries.append(
-                    dict(
-                        values=values,
-                        start=isodate.datetime_isoformat(
-                            measurements[1].datetime.astimezone(start.tzinfo)
-                        ),
-                        duration=isodate.duration_isoformat(
-                            len(values) * measurement_frequency
-                        ),
-                    )
-                )
-                if end_buffer:
-                    timeseries.append(
-                        dict(
-                            value=temp_value,
-                            start=isodate.datetime_isoformat(
-                                measurements[-1].datetime.astimezone(start.tzinfo)
-                            ),
-                            duration=isodate.duration_isoformat(end_buffer),
-                        )
-                    )
-                group.append(dict(connection=asset.name, timeseries=timeseries))
-            else:  # timeseries is univariate
-                group.append(
-                    dict(
-                        connection=asset.name,
-                        values=values,
-                        start=isodate.datetime_isoformat(start),
-                        duration=isodate.duration_isoformat(
-                            len(values) * measurement_frequency
-                        ),
-                    )
-                )
-    if len(group) == 0:
-        return unrecognized_connection_group()
-    elif len(group) == 1:
-        response = dict(unit=unit)
-        response.update(group[0])
-    else:
-        response = dict(group=group, unit=unit)
-    return response
+    return collect_connection_and_value_groups(
+        unit,
+        resolution,
+        connection_groups,
+        horizon,
+        start,
+        duration,
+        preferred_source_ids,
+        fallback_source_ids,
+    )
 
 
 @type_accepted("PostMeterDataRequest")
@@ -241,5 +174,92 @@ def get_service_response(
             response["message"] = invalid_role(requested_access_role)
     else:
         response["services"] = service_listing["services"]
+    d, s = request_processed()
+    return dict(**response, **d), s
+
+
+def collect_connection_and_value_groups(
+    unit: str,
+    resolution: str,
+    connection_groups: List[List[str]],
+    horizon: timedelta,
+    start: datetime_type,
+    duration: timedelta,
+    preferred_source_ids: {
+        Union[int, List[int]]
+    } = None,  # None is interpreted as all sources
+    fallback_source_ids: Union[
+        int, List[int]
+    ] = -1,  # An id = -1 is interpreted as no sources
+) -> Tuple[dict, int]:
+    from flask import current_app
+
+    current_app.logger.info("GETTING")
+
+    from flask import current_app
+
+    # Todo: if resolution is specified, it should be converted properly to a pandas dataframe frequency
+    if resolution == "PT15M":
+        resolution = "15T"
+
+    user_assets = get_assets()
+    if not user_assets:
+        current_app.logger.info("User doesn't seem to have any assets")
+    # user_asset_names = [asset.name for asset in user_assets]
+    user_asset_ids = [asset.id for asset in user_assets]
+
+    end = start + duration
+    value_groups = []
+    new_connection_groups = []  # Each connection in the old connection groups will be interpreted as a separate group
+    for group in connection_groups:
+        asset_names = []
+        for connection in group:
+
+            # Look for the Asset object
+            connection = validate_entity_address(connection)
+            if not connection:
+                current_app.logger.warn(
+                    "Cannot parse this connection's entity address: %s" % connection
+                )
+                return invalid_domain()
+
+            scheme_and_naming_authority, owner_id, asset_id = parse_entity_address(
+                connection
+            )
+            if asset_id in user_asset_ids:
+                asset = Asset.query.filter(Asset.id == asset_id).one_or_none()
+            else:
+                current_app.logger.warn("Cannot identify connection %s" % connection)
+                return unrecognized_connection_group()
+
+            asset_names.append(asset.name)
+        group_response = {"connections": group}
+        ts_df_or_dict = Power.collect(
+            generic_asset_names=asset_names,
+            query_window=(start, end),
+            resolution=resolution,
+            horizon_window=(horizon, horizon),
+            preferred_source_ids=preferred_source_ids,
+            fallback_source_ids=fallback_source_ids,
+            sum_multiple=False,
+        )
+        if isinstance(ts_df_or_dict, dict):
+            for k, v in ts_df_or_dict.items():
+                value_groups.append(v.y.tolist())
+                new_connection_groups.append(k)
+        elif ts_df_or_dict.empty:
+            group_response["values"] = []
+            value_groups.append([])
+            new_connection_groups.append(group)
+        else:
+            group_response["values"] = ts_df_or_dict.y.tolist()
+            value_groups.append(ts_df_or_dict.y.tolist())
+            new_connection_groups.append(group)
+
+    response = groups_to_dict(new_connection_groups, value_groups)
+    response["start"] = isodate.datetime_isoformat(start)
+    response["duration"] = isodate.duration_isoformat(duration)
+    response["unit"] = unit  # TODO: convert to requested unit
+
     d, s = request_processed()
     return dict(**response, **d), s